Consolidate pwm
