--- inodaqv2/command_pwm.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "Arduino.h"
-
-namespace Command
-{
-void command_pwm(const ::String &command);
-}
