+++ inodaqv2/commands.hpp
@@ -6,3 +6,4 @@ void command_aread();
 void command_dig(const ::String &command);
 void command_dread();
 void command_ping();
+void command_pwm(const ::String &command);
